requirements.txt
pyenchant==2.0.0
singledispatch==3.4.0.3
six==1.12.0
